// frontend/src/ts/test/events/types.ts
export type TestEventNoMs =
  | Omit<KeydownEvent, "ms">
  | Omit<KeyupEvent, "ms">
  | Omit<TimerEvent, "ms">
  | InputEventNoMs
  | Omit<CompositionTestEvent, "ms">;

export type InputEventNoMs = Omit<InputEvent, "ms">;

export type TestEventData =
  | KeydownEventData
  | KeyupEventData
  | TimerEventData
  | InputEventData
  | CompositionTestEventData;

export type KeydownEvent = EventProps<"keydown", KeydownEventData>;

export type KeydownEventData = {
  code: Keycode | "NoCode" | `NoCode${number}`;
  ctrl?: true;
  shift?: true;
  alt?: true;
  meta?: true;
};

export type KeyupEvent = EventProps<"keyup", KeyupEventData>;

export type KeyupEventData = {
  code: Keycode | "NoCode" | `NoCode${number}`;
  ctrl?: true;
  shift?: true;
  alt?: true;
  meta?: true;
  estimated?: true; // true if this event never happened, but was estimated (force keyup on test end)
};

export type TimerEvent = EventProps<"timer", TimerEventData>;
// ...
export type CompositionTestEvent = EventProps<
  "composition",
  CompositionTestEventData
>;

export type CompositionTestEventData =
  | {
      event: "start";
      wordIndex: number;
    }
  | {
      event: "update" | "end";
      data: string;
      wordIndex: number;
    };

export type EventLogContext = {
  targetWords: string[];
  // isTimedTest: boolean;
  mode: Config["mode"];
  mode2: ReturnType<typeof getMode2>;
  customTextLimitMode?: CustomTextLimitMode;
  customTextLimitValue?: number;
  isFunboxWithNospacePropertyActive?: boolean;
  bailedOut: boolean;
  koreanStatus: boolean;
};

export const EVENT_LOG_VERSION = 1;
export type EventLog = {
  version: typeof EVENT_LOG_VERSION;
  events: TestEventNoMs[];
  context: EventLogContext;
};